refactor: migrate Input and Password input to typescript
Extract PasswordInput to its own file
