+++ .idea/jsLinters/eslint.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="EslintConfiguration">
-    <extra-options value="--resolve-plugins-relative-to node_modules/@zextras/zapp-cli" />
+    <custom-configuration-file used="true" path="$PROJECT_DIR$/.eslintrc.js" />
   </component>
 </project>
