+++ src/index.ts
@@ -31,7 +31,8 @@ export { default as FileLoader } from './components/inputs/FileLoader';
 export { default as IconButton } from './components/inputs/IconButton';
 export { default as IconCheckbox } from './components/inputs/IconCheckbox';
 export { default as MultiButton } from './components/inputs/MultiButton';
-export { Input, PasswordInput } from './components/inputs/Input';
+export { default as Input } from './components/inputs/Input';
+export { default as PasswordInput } from './components/inputs/PasswordInput';
 export { default as Radio } from './components/inputs/Radio';
 export { default as RadioGroup } from './components/inputs/RadioGroup';
 export { default as SearchInput } from './components/inputs/SearchInput';
